move params parsing to own package
