+++ pkg/action/file.go
@@ -9,6 +9,8 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+
+	"github.com/wndhydrnt/saturn-bot/pkg/params"
 )
 
 func createContentReader(taskPath, value string) (io.ReadCloser, error) {
@@ -33,7 +35,7 @@ func createContentReader(taskPath, value string) (io.ReadCloser, error) {
 
 type FileCreateFactory struct{}
 
-func (f FileCreateFactory) Create(params Params, taskPath string) (Action, error) {
+func (f FileCreateFactory) Create(params params.Params, taskPath string) (Action, error) {
 	var content string
 	if params["content"] != nil {
 		contentCast, ok := params["content"].(string)
@@ -175,7 +177,7 @@ func (a *fileCreate) String() string {
 
 type FileDeleteFactory struct{}
 
-func (f FileDeleteFactory) Create(params Params, _ string) (Action, error) {
+func (f FileDeleteFactory) Create(params params.Params, _ string) (Action, error) {
 	if params["path"] == nil {
 		return nil, fmt.Errorf("required parameter `path` not set")
 	}
